cloud: fix global average calculation

+++ src/main/java/org/jlab/clas/std/orchestrators/AbstractOrchestrator.java
@@ -173,7 +173,7 @@ abstract class AbstractOrchestrator {
             return sum;
         }
 
-        long globalTime() {
+        double globalTime() {
             return endTime.get() - startTime.get();
         }
